+++ lib/src/http_client_none.dart
@@ -5,7 +5,7 @@ import 'http_client.dart';
 class HttpClientRequesterNone extends HttpClientRequester {
 
   @override
-  Future<HttpResponse> doHttpRequest(HttpClient client, HttpRequest request) {
+  Future<HttpResponse> doHttpRequest(HttpClient client, HttpRequest request, bool log ) {
     return Future.error( HttpError( request.url, request.requestURL, 0, 'No HttpClientRequester for ${ request.method } request: ${ request.requestURL }', null) ) ;
   }